+++ RobotRaceGradle/src/main/java/bodies/assembly/Vertex.java
@@ -39,29 +39,29 @@ public class Vertex {
     /**
      * The position of this Vertex in x, y and z components.
      */
-    private double positionX;
-    private double positionY;
-    private double positionZ;
+    private float positionX;
+    private float positionY;
+    private float positionZ;
     /**
      * The normal of this Vertex (mainly useful when using surfaces) in x, y and
      * z components.
      */
-    private double normalX;
-    private double normalY;
-    private double normalZ;
+    private float normalX;
+    private float normalY;
+    private float normalZ;
     /**
      * The texture coordinates of this Vertex (mainly useful when using
      * surfaces) in x, y and z components.
      */
-    private double textureX;
-    private double textureY;
-    private double textureZ;
+    private float textureX;
+    private float textureY;
+    private float textureZ;
 
     /**
      * Constructor.
      */
     public Vertex() {
-        this(0d, 0d, 0d);
+        this(0f, 0f, 0f);
     }
 
     /**
@@ -71,8 +71,8 @@ public class Vertex {
      * @param positionY Component of the position in the Y axis.
      * @param positionZ Component of the position in the Z axis.
      */
-    public Vertex(double positionX, double positionY, double positionZ) {
-        this(positionX, positionY, positionZ, 0d, 0d, 0d);
+    public Vertex(float positionX, float positionY, float positionZ) {
+        this(positionX, positionY, positionZ, 0f, 0f, 0f);
     }
 
     /**
@@ -85,8 +85,8 @@ public class Vertex {
      * @param normalY   Component of the normal in the Y axis.
      * @param normalZ   Component of the normal in the Z axis.
      */
-    public Vertex(double positionX, double positionY, double positionZ, double normalX, double normalY, double normalZ) {
-        this(positionX, positionY, positionZ, 0d, 0d, 0d, 0d, 0d, 0d);
+    public Vertex(float positionX, float positionY, float positionZ, float normalX, float normalY, float normalZ) {
+        this(positionX, positionY, positionZ, 0f, 0f, 0f, 0f, 0f, 0f);
     }
 
     /**
@@ -102,7 +102,7 @@ public class Vertex {
      * @param textureY  Component of the texture in the Y axis.
      * @param textureZ  Component of the texture in the Z axis.
      */
-    public Vertex(double positionX, double positionY, double positionZ, double normalX, double normalY, double normalZ, double textureX, double textureY, double textureZ) {
+    public Vertex(float positionX, float positionY, float positionZ, float normalX, float normalY, float normalZ, float textureX, float textureY, float textureZ) {
         setPositionC(positionX, positionY, positionZ);
         setNormalC(normalX, normalY, normalZ);
         setTextureC(textureX, textureY, textureZ);
@@ -113,8 +113,8 @@ public class Vertex {
      *
      * @param position The position.
      */
-    public Vertex(double[] position) {
-        this(position, new double[]{0d, 0d, 0d}, new double[]{0d, 0d, 0d});
+    public Vertex(float[] position) {
+        this(position, new float[]{0f, 0f, 0f}, new float[]{0f,0f,0f});
     }
 
     /**
@@ -123,8 +123,8 @@ public class Vertex {
      * @param position The position.
      * @param normal   The normal.
      */
-    public Vertex(double[] position, double[] normal) {
-        this(position, normal, new double[]{0d, 0d, 0d});
+    public Vertex(float[] position, float[] normal) {
+        this(position, normal, new float[]{0f, 0f, 0f});
     }
 
     /**
@@ -134,7 +134,7 @@ public class Vertex {
      * @param normal   The normal.
      * @param texture  The texture.
      */
-    public Vertex(double[] position, double[] normal, double[] texture) {
+    public Vertex(float[] position, float[] normal, float[] texture) {
         this(position[IND_X], position[IND_Y], position[IND_Z], normal[IND_X], normal[IND_Y], normal[IND_Z], texture[IND_X], texture[IND_Y], texture[IND_Z]);
     }
 
@@ -144,7 +144,7 @@ public class Vertex {
      * @param position The position.
      */
     public Vertex(Vector position) {
-        this(position.x(), position.y(), position.z());
+        this((float) position.x(), (float) position.y(), (float) position.z());
     }
 
     /**
@@ -154,7 +154,7 @@ public class Vertex {
      * @param normal   The normal.
      */
     public Vertex(Vector position, Vector normal) {
-        this(position.x(), position.y(), position.z(), normal.x(), normal.y(), normal.z());
+        this((float) position.x(), (float) position.y(), (float) position.z(), (float) normal.x(), (float) normal.y(), (float) normal.z());
     }
 
     /**
@@ -165,7 +165,7 @@ public class Vertex {
      * @param texture  The texture.
      */
     public Vertex(Vector position, Vector normal, Vector texture) {
-        this(position.x(), position.y(), position.z(), normal.x(), normal.y(), normal.z(), texture.x(), texture.y(), texture.z());
+        this((float) position.x(), (float) position.y(), (float) position.z(), (float) normal.x(),(float)  normal.y(), (float) normal.z(), (float) texture.x(), (float) texture.y(), (float) texture.z());
     }
 
     /**
@@ -175,7 +175,7 @@ public class Vertex {
      * @param positionY Component of the new position in the Y axis.
      * @param positionZ Component of the new position in the Z axis.
      */
-    public final void setPositionC(double positionX, double positionY, double positionZ) {
+    public final void setPositionC(float positionX, float positionY, float positionZ) {
         this.positionX = positionX;
         this.positionY = positionY;
         this.positionZ = positionZ;
@@ -186,7 +186,7 @@ public class Vertex {
      *
      * @param position The new position.
      */
-    public final void setPositionA(double[] position) {
+    public final void setPositionA(float[] position) {
         this.setPositionC(position[IND_X], position[IND_Y], position[IND_Z]);
     }
 
@@ -196,7 +196,7 @@ public class Vertex {
      * @param position The new position.
      */
     public final void setPositionV(Vector position) {
-        this.setPositionC(position.x(), position.y(), position.z());
+        this.setPositionC((float) position.x(), (float) position.y(), (float) position.z());
     }
 
     /**
@@ -204,8 +204,8 @@ public class Vertex {
      *
      * @return The position of this Vertex.
      */
-    public final double[] getPositionA() {
-        double[] position = new double[COORD_COUNT];
+    public final float[] getPositionA() {
+        float[] position = new float[COORD_COUNT];
         position[IND_X] = positionX;
         position[IND_Y] = positionY;
         position[IND_Z] = positionZ;
@@ -228,7 +228,7 @@ public class Vertex {
      * @param normalY Component of the new normal in the Y axis.
      * @param normalZ Component of the new normal in the Z axis.
      */
-    public final void setNormalC(double normalX, double normalY, double normalZ) {
+    public final void setNormalC(float normalX, float normalY, float normalZ) {
         this.normalX = normalX;
         this.normalY = normalY;
         this.normalZ = normalZ;
@@ -239,7 +239,7 @@ public class Vertex {
      *
      * @param normal The new normal of this Vertex.
      */
-    public final void setNormalA(double[] normal) {
+    public final void setNormalA(float[] normal) {
         this.setNormalC(normal[IND_X], normal[IND_Y], normal[IND_Z]);
     }
 
@@ -249,7 +249,7 @@ public class Vertex {
      * @param normal The new normal of this Vertex.
      */
     public final void setNormalV(Vector normal) {
-        this.setNormalC(normal.x(), normal.y(), normal.z());
+        this.setNormalC((float) normal.x(), (float) normal.y(), (float) normal.z());
     }
 
     /**
@@ -257,8 +257,8 @@ public class Vertex {
      *
      * @return The normal of this Vertex.
      */
-    public final double[] getNormalA() {
-        double[] normal = new double[COORD_COUNT];
+    public final float[] getNormalA() {
+        float[] normal = new float[COORD_COUNT];
         normal[IND_X] = normalX;
         normal[IND_Y] = normalY;
         normal[IND_Z] = normalZ;
@@ -281,7 +281,7 @@ public class Vertex {
      * @param textureY Component of the new texture in the Y axis.
      * @param textureZ Component of the new texture in the Z axis.
      */
-    public final void setTextureC(double textureX, double textureY, double textureZ) {
+    public final void setTextureC(float textureX, float textureY, float textureZ) {
         this.textureX = textureX;
         this.textureY = textureY;
         this.textureZ = textureZ;
@@ -292,7 +292,7 @@ public class Vertex {
      *
      * @param texture The new texture of this Vertex.
      */
-    public final void setTextureA(double[] texture) {
+    public final void setTextureA(float[] texture) {
         this.setTextureC(texture[IND_X], texture[IND_Y], texture[IND_Z]);
     }
 
@@ -302,7 +302,7 @@ public class Vertex {
      * @param texture The new texture of this Vertex.
      */
     public final void setTextureV(Vector texture) {
-        this.setTextureC(texture.x(), texture.y(), texture.z());
+        this.setTextureC((float) texture.x(), (float) texture.y(), (float) texture.z());
     }
 
     /**
@@ -310,8 +310,8 @@ public class Vertex {
      *
      * @return The texture of this Vertex.
      */
-    public final double[] getTextureA() {
-        double[] texture = new double[COORD_COUNT];
+    public final float[] getTextureA() {
+        float[] texture = new float[COORD_COUNT];
         texture[IND_X] = textureX;
         texture[IND_Y] = textureY;
         texture[IND_Z] = textureZ;
